Changed Id and FKID from int to long long int. Added sqlCmd.SetCommandText() to the Get functions.

+++ DbConnection.h
@@ -11,16 +11,16 @@ public:
 	
 	struct Book
 	{
-		int Id;
+		long long int Id;
 		std::string Title;
 		std::string ISBN;
 		std::string PublishDate;
-		int AuthorId;
+		long long int AuthorId;
 	};
 
 	struct Author
 	{
-		int Id;
+		long long int Id;
 		std::string FirstName;
 		std::string LastName;
 	};
